+++ src/component/association-select.js
@@ -1,14 +1,17 @@
-import {logger} from '../aurelia-orm';
-import getProp from 'get-prop';
-import {inject} from 'aurelia-dependency-injection';
-import {bindingMode, BindingEngine} from 'aurelia-binding';
-import {bindable, customElement} from 'aurelia-templating';
-import {EntityManager, Entity, OrmMetadata} from '../aurelia-orm';
+import getProp from "get-prop";
+import {inject} from "aurelia-dependency-injection";
+import {bindingMode, BindingEngine} from "aurelia-binding";
+import {bindable, customElement} from "aurelia-templating";
+import {logger, EntityManager, Entity, OrmMetadata} from "../aurelia-orm";
+import {resolvedView} from 'aurelia-view-manager';
 
 @customElement('association-select')
+@resolvedView('spoonx/orm', 'association-select')
 @inject(BindingEngine, EntityManager, Element)
 export class AssociationSelect {
   @bindable criteria;
+
+  @bindable name = '';
 
   @bindable repository;
 
@@ -34,6 +37,10 @@ export class AssociationSelect {
 
   @bindable selectablePlaceholder = false;
 
+  @bindable placeholderValue = 0;
+
+  @bindable disabled = false;
+
   @bindable placeholderText;
 
   ownMeta;
@@ -53,14 +60,15 @@ export class AssociationSelect {
   /**
    * (Re)Load the data for the select.
    *
-   * @param {string|Array|Object} [reservedValue]
+   * @param {string|Array|{}} [reservedValue]
    *
    * @return {Promise}
    */
   load(reservedValue) {
     return this.buildFind()
       .then(options => {
-        let result   = options;
+        let result = options;
+
         this.options = Array.isArray(result) ? result : [result];
 
         this.setValue(reservedValue);
@@ -70,7 +78,7 @@ export class AssociationSelect {
   /**
    * Set the value for the select.
    *
-   * @param {string|Array|Object} value
+   * @param {string|Array|{}} value
    */
   setValue(value) {
     if (!value) {
@@ -112,20 +120,21 @@ export class AssociationSelect {
    * @return {Promise}
    */
   buildFind() {
-    let repository    = this.repository;
-    let criteria      = this.getCriteria();
-    let findPath      = repository.getResource();
+    let repository = this.repository;
+    let criteria   = this.getCriteria();
+    let findPath   = repository.getResource();
+
     criteria.populate = false;
 
     // Check if there are `many` associations. If so, the repository find path changes.
     // the path will become `/:association/:id/:entity`.
     if (this.manyAssociation) {
-      let assoc = this.manyAssociation;
+      let manyAssociation = this.manyAssociation;
 
       // When disabling populate here, the API won't return any data.
       delete criteria.populate;
 
-      findPath = `${assoc.getResource()}/${assoc.getId()}/${findPath}`;
+      findPath = `${manyAssociation.resource}/${manyAssociation.entity.getId()}/${manyAssociation.property}`;
     } else if (this.association) {
       let associations = Array.isArray(this.association) ? this.association : [this.association];
 
@@ -134,7 +143,12 @@ export class AssociationSelect {
       });
     }
 
-    return repository.findPath(findPath, criteria).catch(error => this.error = error);
+    return repository.findPath(findPath, criteria)
+      .catch(error => {
+        this.error = error;
+
+        return error;
+      });
   }
 
   /**
@@ -144,7 +158,7 @@ export class AssociationSelect {
    */
   verifyAssociationValues() {
     if (this.manyAssociation) {
-      return !!this.manyAssociation.getId();
+      return !!this.manyAssociation.entity.getId();
     }
 
     if (this.association) {
@@ -186,21 +200,22 @@ export class AssociationSelect {
   }
 
   /**
-   * Check if the value is changed
+   * Check if the element property has changed
    *
-   * @param  {string|{}}   newVal New value
-   * @param  {[string|{}]} oldVal Old value
-   * @return {Boolean}     Whenever the value is changed
+   * @param  {string}      property
+   * @param  {string|{}}   newVal
+   * @param  {string|{}}   oldVal
+   *
+   * @return {boolean}
    */
   isChanged(property, newVal, oldVal) {
     return !this[property] || !newVal || (newVal === oldVal);
   }
 
   /**
-   * Change resource
+   * Changed resource handler
    *
-   * @param  {{}} newVal New criteria value
-   * @param  {{}} oldVal Old criteria value
+   * @param  {string} resource
    */
   resourceChanged(resource) {
     if (!resource) {
@@ -211,10 +226,10 @@ export class AssociationSelect {
   }
 
   /**
-   * Change criteria
+   * Changed criteria handler
    *
-   * @param  {{}} newVal New criteria value
-   * @param  {{}} oldVal Old criteria value
+   * @param  {{}} newVal
+   * @param  {{}} oldVal
    */
   criteriaChanged(newVal, oldVal) {
     if (this.isChanged('criteria', newVal, oldVal)) {
@@ -226,10 +241,9 @@ export class AssociationSelect {
     }
   }
 
-  /**
-   * When attached to the DOM, initialize the component.
-   */
-  attached() {
+  bind() {
+    this.resourceChanged(this.resource);
+
     if (!this.association && !this.manyAssociation) {
       this.load(this.value);
 
@@ -239,7 +253,28 @@ export class AssociationSelect {
     this.ownMeta = OrmMetadata.forTarget(this.entityManager.resolveEntityReference(this.repository.getResource()));
 
     if (this.manyAssociation) {
-      this.observe(this.manyAssociation);
+      if (this.manyAssociation instanceof Entity) {
+        this.manyAssociation = {entity: this.manyAssociation};
+      }
+
+      let manyAssociation = this.manyAssociation;
+
+      if (!manyAssociation.entity) {
+        throw new Error(
+          'Invalid value provided for many-association. '
+          + 'Expected instance of Entity, or object literal {entity, property}.'
+        );
+      }
+
+      let manyEntity = manyAssociation.entity;
+
+      if (!manyAssociation.property) {
+        manyAssociation.property = this.propertyForResource(manyEntity.getMeta(), this.repository.getResource());
+      }
+
+      manyAssociation.resource = manyEntity.getResource();
+
+      this.observe(manyEntity);
     }
 
     if (this.association) {