+++ src/orm-metadata.js
@@ -20,11 +20,12 @@ export class Metadata {
    */
   constructor() {
     this.metadata = {
-      repository: DefaultRepository,
-      resource: null,
-      endpoint: null,
-      name: null,
-      idProperty: 'id',
+      repository  : DefaultRepository,
+      identifier  : null,
+      resource    : null,
+      endpoint    : null,
+      name        : null,
+      idProperty  : 'id',
       associations: {}
     };
   }
@@ -35,7 +36,7 @@ export class Metadata {
    * @param {string} key
    * @param {*} value
    *
-   * @return {Metadata} this
+   * @return {Metadata} itself
    * @chainable
 */
   addTo(key, value) {
@@ -57,7 +58,7 @@ export class Metadata {
    * @param {string|*} valueOrNestedKey
    * @param {null|*} [valueOrNull]
    *
-   * @return {Metadata} this
+   * @return {Metadata} itself
    * @chainable
    */
   put(key, valueOrNestedKey, valueOrNull) {