+++ src/repository.js
@@ -40,7 +40,7 @@ export class Repository {
   /**
    * Set the associated entity's meta data
    *
-   * @param {Object} meta
+   * @param {{}} meta
    */
   setMeta(meta) {
     this.meta = meta;
@@ -48,10 +48,32 @@ export class Repository {
 
   /**
    * Get the associated entity's meta data.
-   * @return {Object}
+   * @return {{}}
    */
   getMeta() {
     return this.meta;
+  }
+
+  /**
+   * Set the identifier
+   *
+   * @param {string} identifier
+   * @return {Repository} this
+   * @chainable
+   */
+  setIdentifier(identifier) {
+    this.identifier = identifier;
+
+    return this;
+  }
+
+  /**
+   * Get the identifier
+   *
+   * @return {string|null}
+   */
+  getIdentifier() {
+    return this.identifier;
   }
 
   /**
@@ -79,7 +101,7 @@ export class Repository {
   /**
    * Perform a find query and populate entities with the retrieved data.
    *
-   * @param {{}|Number|String} criteria Criteria to add to the query. A plain String or Number will be used as relative path.
+   * @param {{}|number|string} criteria Criteria to add to the query. A plain string or number will be used as relative path.
    * @param {boolean}          [raw]    Set to true to get a POJO in stead of populated entities.
    *
    * @return {Promise<Entity|[Entity]>}
@@ -89,16 +111,40 @@ export class Repository {
   }
 
   /**
-   * Perform a find query for `path` and populate entities with the retrieved data.
+   * Perform a find query and populate entities with the retrieved data, limited to one result.
    *
-   * @param {string}           path
-   * @param {{}|Number|String} criteria Criteria to add to the query. A plain String or Number will be used as relative path.
+   * @param {{}|number|string} criteria Criteria to add to the query. A plain string or number will be used as relative path.
    * @param {boolean}          [raw]    Set to true to get a POJO in stead of populated entities.
    *
    * @return {Promise<Entity|[Entity]>}
    */
-  findPath(path, criteria, raw) {
-    let findQuery = this.getTransport().find(path, criteria);
+  findOne(criteria, raw) {
+    return this.findPath(this.resource, criteria, raw, true);
+  }
+
+  /**
+   * Perform a find query for `path` and populate entities with the retrieved data.
+   *
+   * @param {string}           path
+   * @param {{}|number|string} criteria Criteria to add to the query. A plain string or number will be used as relative path.
+   * @param {boolean}          [raw]    Set to true to get a POJO in stead of populated entities.
+   * @param {boolean}          [single] Whether or not this is a findOne.
+   *
+   * @return {Promise<Entity|[Entity]>}
+   */
+  findPath(path, criteria, raw, single) {
+    let transport = this.getTransport();
+    let findQuery;
+
+    if (single) {
+      if (typeof criteria === 'object' && criteria !== null) {
+        criteria.limit = 1;
+      }
+
+      findQuery = transport.findOne(path, criteria);
+    } else {
+      findQuery = transport.find(path, criteria);
+    }
 
     if (raw) {
       return findQuery;
@@ -128,7 +174,7 @@ export class Repository {
    *
    * @param {null|{}} criteria
    *
-   * @return {Promise<Number>}
+   * @return {Promise<number>}
    */
   count(criteria) {
     return this.getTransport().find(this.resource + '/count', criteria);
@@ -138,22 +184,23 @@ export class Repository {
    * Get new populated entity or entities based on supplied data including associations
    *
    * @param {{}|[{}]} data|[data] The data to populate with
+   * @param {boolean} [clean]     Mark the entities as clean or not
    *
    * @return {Entity|[Entity]}
    */
-  populateEntities(data) {
+  populateEntities(data, clean) {
     if (!data) {
       return null;
     }
 
     if (!Array.isArray(data)) {
-      return this.getPopulatedEntity(data);
+      return this.getPopulatedEntity(data, null, clean);
     }
 
     let collection = [];
 
     data.forEach(source => {
-      collection.push(this.getPopulatedEntity(source));
+      collection.push(this.getPopulatedEntity(source, null, clean));
     });
 
     return collection;
@@ -162,12 +209,13 @@ export class Repository {
   /**
    * Populate a (new) entity including associations
    *
-   * @param {{}}     data The data to populate with
-   * @param {Entity} [entity] optional. if not set, a new entity is returned
+   * @param {{}}      data     The data to populate with
+   * @param {Entity}  [entity] optional. if not set, a new entity is returned
+   * @param {boolean} [clean]  Mark the entities as clean or not
    *
    * @return {Entity}
    */
-  getPopulatedEntity(data, entity) {
+  getPopulatedEntity(data, entity, clean) {
     entity             = entity || this.getNewEntity();
     let entityMetadata = entity.getMeta();
     let populatedData  = {};
@@ -181,7 +229,13 @@ export class Repository {
       let value = data[key];
 
       if (entityMetadata.has('types', key)) {
-        populatedData[key] = typer.cast(value, entityMetadata.fetch('types', key));
+        const dataType = entityMetadata.fetch('types', key);
+
+        if ((dataType === 'date' || dataType === 'datetime') && !value) {
+          continue;
+        }
+
+        populatedData[key] = typer.cast(value, dataType);
 
         continue;
       }
@@ -193,11 +247,12 @@ export class Repository {
         continue;
       }
 
-      let repository     = this.entityManager.getRepository(entityMetadata.fetch('associations', key).entity);
-      populatedData[key] = repository.populateEntities(value);
+      let repository = this.entityManager.getRepository(entityMetadata.fetch('associations', key).entity);
+
+      populatedData[key] = repository.populateEntities(value, clean);
     }
 
-    return entity.setData(populatedData);
+    return entity.setData(populatedData, clean);
   }
 
   /**
@@ -206,7 +261,7 @@ export class Repository {
    * @return {Entity}
    */
   getNewEntity() {
-    return this.entityManager.getEntity(this.resource);
+    return this.entityManager.getEntity(this.identifier || this.resource);
   }
 
   /**
@@ -219,13 +274,15 @@ export class Repository {
     let associations = entity.getMeta().fetch('associations');
 
     for (let property in associations) {
-      let assocMeta = associations[property];
+      if (associations.hasOwnProperty(property)) {
+        let assocMeta = associations[property];
 
-      if (assocMeta.type !== 'entity') {
-        continue;
+        if (assocMeta.type !== 'entity') {
+          continue;
+        }
+
+        entity[property] = this.entityManager.getRepository(assocMeta.entity).getNewEntity();
       }
-
-      entity[property] = this.entityManager.getRepository(assocMeta.entity).getNewEntity();
     }
 
     return entity;